feat: Update to a model-agnostic model
The previous version was too tightly integrated with the SET, when the
set changed the format, a major refactor was needed, this new version
uses a independent format that will be easier to migrate.

The VersionMigrator is the main class that is in charge of the migration
of the v1 data (the same format that uses the aranduka) to the new
format. In a future this migrator will be in charge of migrating any
version to the latest one and the entire system will only use the latest
one.

Signed-off-by: Arturo Volpe <arturovolpe@gmail.com>